WIP: Interior point test.
Based on quickly selecting the triangles that intersect with the
Z-plane of the point, then performing ray-casting among those
to know how many intersection we have. Should be okish with
performance.

I reused to trick of shifting the Z positions by a small fixed quantity
separately for the points and the triangles.

Does not work perfectly yet. I think I forgot to deal with cases
where the intersection lies exactly on an edge, which cannot be
avoided. This is why the JUnit test fails in one case.