Better test for interior point.
The hollow cube test does not pass, despite of the assertion of
inside being correct. When we extrude the inside hollow cube, the
mesh interpolates triangles and yield triangles that have a 45º angle.
Because we validate the interior point test against a predicate testing
whether a point is inside the intervals that were used to create the
image on which we created the mesh, there are small spaces near
the inside limit where a point can be correctly inside the mesh
and also inside the extruded cube, making this test fail.