Make a JUnit test for interior test that fails on an edge case.
Test the situation where the ray casting algorithm will just cross an
edge between two triangles with opposite normals. It makes the test
fail in this case, because the ray casting treats the intersection
with the 2 triangles as one.
They should be treated as two if their normals projected on
the ray are opposed.