Deal with interior test edge case properly.
Using the normals to determine whether we cross the mesh border or
not, when we have identical X intersections. We don't have to deal
with cases where we cross a vertex, because we don't: in this
algorithm, the vertices Z position is shifted to odd intervals, while
the points Z position is shifted to even intervals.